requirements.txt
tqdm==4.66.2
transformers==4.40.0
torch==2.3.1
torchvision
timm==1.0.8
PyMuPDF==1.24.9
numpy==1.26.4
pillow==10.4.0
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1
protobuf==4.23.4
opencv-python==4.6.0.66
# Detectron2 from GitHub
git+https://github.com/facebookresearch/detectron2.git